+++ target-gen/src/commands/test.rs
@@ -95,7 +95,9 @@ pub fn cmd_test(
     // TODO: The sector used here is not necessarily the sector the flash algorithm targets.
     // Make this configurable.
     let mut readback = vec![0; flash_properties.sectors[0].size as usize];
-    session.core(0)?.read_8(0x0, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start, &mut readback)?;
     assert!(
         !readback.iter().any(|v| *v != erased_state),
         "Not all sectors were erased"
@@ -106,10 +108,12 @@ pub fn cmd_test(
         .into_iter()
         .map(|n| (n % 256) as u8)
         .collect::<Vec<_>>();
-    loader.add_data(0x1, &data)?;
+    loader.add_data(flash_properties.address_range.start + 1, &data)?;
     run_flash_download(&mut session, loader, progress.clone(), true)?;
     let mut readback = vec![0; data_size as usize];
-    session.core(0)?.read_8(0x1, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start + 1, &mut readback)?;
     assert_eq!(readback, data);
 
     println!("{test}: Erasing the entire chip and writing two pages ...");
@@ -117,7 +121,9 @@ pub fn cmd_test(
     // TODO: The sector used here is not necessarily the sector the flash algorithm targets.
     // Make this configurable.
     let mut readback = vec![0; flash_properties.sectors[0].size as usize];
-    session.core(0)?.read_8(0x0, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start, &mut readback)?;
     assert!(
         !readback.iter().any(|v| *v != erased_state),
         "Not all sectors were erased"
@@ -128,10 +134,12 @@ pub fn cmd_test(
         .into_iter()
         .map(|n| (n % 256) as u8)
         .collect::<Vec<_>>();
-    loader.add_data(0x1, &data)?;
+    loader.add_data(flash_properties.address_range.start + 1, &data)?;
     run_flash_download(&mut session, loader, progress.clone(), true)?;
     let mut readback = vec![0; data_size as usize];
-    session.core(0)?.read_8(0x1, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start + 1, &mut readback)?;
     assert_eq!(readback, data);
 
     println!("{test}: Erasing sectorwise and writing two pages double buffered ...");
@@ -139,7 +147,9 @@ pub fn cmd_test(
     // TODO: The sector used here is not necessarily the sector the flash algorithm targets.
     // Make this configurable.
     let mut readback = vec![0; flash_properties.sectors[0].size as usize];
-    session.core(0)?.read_8(0x0, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start, &mut readback)?;
     assert!(
         !readback.iter().any(|v| *v != erased_state),
         "Not all sectors were erased"
@@ -150,10 +160,12 @@ pub fn cmd_test(
         .into_iter()
         .map(|n| (n % 256) as u8)
         .collect::<Vec<_>>();
-    loader.add_data(0x1, &data)?;
+    loader.add_data(flash_properties.address_range.start + 1, &data)?;
     run_flash_download(&mut session, loader, progress, false)?;
     let mut readback = vec![0; data_size as usize];
-    session.core(0)?.read_8(0x1, &mut readback)?;
+    session
+        .core(0)?
+        .read_8(flash_properties.address_range.start + 1, &mut readback)?;
     assert_eq!(readback, data);
 
     Ok(())